// web/src/generated/models/ApiV1ClusterConnection.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ApiV1ClusterConnection
 */
export interface ApiV1ClusterConnection {
    /**
     *
     * @type {string}
     * @memberof ApiV1ClusterConnection
     */
    address: string;
    /**
     *
     * @type {string}
     * @memberof ApiV1ClusterConnection
     */
    ca?: string;
    /**
     *
     * @type {string}
     * @memberof ApiV1ClusterConnection
     */
    type: ApiV1ClusterConnectionTypeEnum;
    /**
     *
     * @type {string}
     * @memberof ApiV1ClusterConnection
     */
    cert?: string;
    /**
     *
     * @type {string}
     * @memberof ApiV1ClusterConnection
     */
    key?: string;
    /**
     *
     * @type {string}
     * @memberof ApiV1ClusterConnection
     */
    token?: string;
}

/**
 * @export
 */
export const ApiV1ClusterConnectionTypeEnum = {
    Unspecified: "UNSPECIFIED",
    None: "NONE",
    TlsCert: "TLS_CERT",
    TlsToken: "TLS_TOKEN",
} as const;
export type ApiV1ClusterConnectionTypeEnum =
    (typeof ApiV1ClusterConnectionTypeEnum)[keyof typeof ApiV1ClusterConnectionTypeEnum];

/**
 * Check if a given object implements the ApiV1ClusterConnection interface.
 */
export function instanceOfApiV1ClusterConnection(value: object): value is ApiV1ClusterConnection {
    if (!("address" in value) || value["address"] === undefined) return false;
    if (!("type" in value) || value["type"] === undefined) return false;
    return true;
}

export function ApiV1ClusterConnectionFromJSON(json: any): ApiV1ClusterConnection {
    return ApiV1ClusterConnectionFromJSONTyped(json, false);
}

export function ApiV1ClusterConnectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiV1ClusterConnection {
    if (json == null) {
        return json;
    }
    return {
        address: json["address"],
        ca: json["ca"] == null ? undefined : json["ca"],
        type: json["type"],
        cert: json["cert"] == null ? undefined : json["cert"],
        key: json["key"] == null ? undefined : json["key"],
        token: json["token"] == null ? undefined : json["token"],
    };
}

export function ApiV1ClusterConnectionToJSON(json: any): ApiV1ClusterConnection {
    return ApiV1ClusterConnectionToJSONTyped(json, false);
}

export function ApiV1ClusterConnectionToJSONTyped(
    value?: ApiV1ClusterConnection | null,
    ignoreDiscriminator: boolean = false,
): any {
    if (value == null) {
        return value;
    }

    return {
        address: value["address"],
        ca: value["ca"],
        type: value["type"],
        cert: value["cert"],
        key: value["key"],
        token: value["token"],
    };
}